refactor: centralize db pool
