+++ lib/susanin.js
@@ -1,26 +1,9 @@
 import { getFlights, getAirportByIata } from './data.js';
 import { DateTime } from 'luxon';
-import mysql from 'mysql2/promise';
-import { start } from 'node:repl';
+import pool from './db.js';
 
 const flightsFromCache = new Map();
 
-const { DB_HOST, DB_PORT, DB_NAME, DB_USER, DB_PASS } = process.env;
-
-if (!DB_HOST || !DB_PORT || !DB_NAME || !DB_USER || !DB_PASS) {
-  console.error('Missing database configuration. Set DB_HOST, DB_PORT, DB_NAME, DB_USER and DB_PASS.');
-  process.exit(1);
-}
-
-const pool = mysql.createPool({
-  host: DB_HOST,
-  port: DB_PORT,
-  user: DB_USER,
-  password: DB_PASS,
-  database: DB_NAME,
-  timezone: 'Z',
-  dateStrings: true,
-});
 
 class Route {
   constructor(legs) {
